+++ packages/fcl/src/config-utils.js
@@ -1,5 +1,6 @@
 import {config} from "@onflow/sdk"
-import {invariant} from "@onflow/util-invariant"
+import {constructApiQueryParams} from "./discovery/services"
+import {VERSION} from "./VERSION"
 
 const isServerSide = () => typeof window === "undefined"
 
@@ -32,10 +33,14 @@ export async function configLens(regex) {
 }
 
 export async function buildAuthnConfig() {
-  const discoveryWallet = await config.first([
+  const discoveryWalletUrl = await config.first([
     "discovery.wallet",
     "challenge.handshake",
   ])
+
+  const include = await config.get("discovery.authn.include", [])
+  const queryParams = constructApiQueryParams({version: VERSION, include})
+  const discoveryWallet = `${discoveryWalletUrl}${queryParams}`
 
   const discoveryWalletMethod = await config.first([
     "discovery.wallet.method",