+++ packages/fcl/CHANGELOG.md
@@ -1,6 +1,19 @@
 ## Unreleased
 
 - 2021-12-01 -- Internal: Wrap authz in resolve to dedupe accounts. Remove user `notExpired` check.
+- 2021-11-30 -- Allow apps to add opt-in wallets in Discovery with config.
+
+```javascript
+import { config } from "@onflow/fcl"
+
+// Include supports discovery.wallet or discovery.authn.endpoint
+config({
+  "discovery.wallet": "https://fcl-discovery.onflow.org/testnet/authn",
+  "discovery.authn.endpoint": "https://fcl-discovery.onflow.org/api/testnet/authn",
+  "discovery.authn.include": ["0x123"] // Service account address
+})
+```
+
 - 2021-11-22 -- Updates to use account key `HashAlgorithm` in `verifyUSerSignatures` instead of default `SHA3_256`
 - 2021-11-19 -- Update `buildUser` with with correct data prop for `expiresAt`